LayoutTests/fast/forms/ios/file-upload-panel-dismiss-when-view-removed-from-window-expected.txt

Test that the file upload panel is dismissed when its presenting view is removed from the window on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Present file picker
PASS areArraysEqual(typeIdentifiers, ['public.plain-text']) is true

Remove view from window
PASS areArraysEqual(typeIdentifiers, []) is true
PASS successfullyParsed is true

TEST COMPLETE
